qcs8550-common: ueventd: Modify USB root hub authorize setting
Change the ownership of the authorize file to allow the USB bus reset audio
recovery mechanism in the USB HAL to work.  This only modifies the
ownership of the USB root hubs that are present on the system.  For all
other connected devices, the default ownership is fine, as the recovery is
only allowed on direct connection to the USB root hub.

Change-Id: Idbce78b52054f75c09891aa1c1c827dda60f1d0a